LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/data-transfer-files.https-expected.txt




FAIL Posting a File in a navigation handled by a service worker assert_equals: expected "foobar" but got "FieldStorage('testinput', 'name', 'foobar')"
